refactor: move model definition to data
