+++ src/data/classicSensor.ts
@@ -1,4 +1,4 @@
-import { isCasesSignal, isCountSignal, isDeathSignal, isPropSignal } from '../data';
+import { isCasesSignal, isCountSignal, isDeathSignal, isPropSignal } from './signals';
 import { ensureSensorStructure, Sensor } from './sensor';
 
 export interface CasesOrDeathOptions {
